requirements.txt
clarifai
requests
Pillow
jsonschema
future
configparser

